requirements.txt
speedtest-cli
numpy
pandas
matplotlib
psutil
